% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeperiods.R
\name{mergePeriods}
\alias{mergePeriods}
\title{Merge dispensations so that the implied treatment periods are not overlapping.}
\usage{
mergePeriods(data, dates, doses, cumulative = F)
}
\arguments{
\item{data}{The data frame.}

\item{dates}{The column name containing the dates of treatment initiations.}

\item{doses}{The column name containing the treatment length.}

\item{cumilative}{Indicate swhether the days in overlapping treatment periods should accumulate.}
}
\value{
A data frame without overlapping periods.
}
\description{
Merge dispensations so that the implied treatment periods are not overlapping.
}
\details{
Given data with a column of dates ("date") as a date variable, a column of quantity dispensed ("quantity"),
mergePeriods(data, "date", "quantity", FALSE) returns an aggregation of data where overlapping periods
have been merged to a single period.
}
\examples{
mergePeriods()
}
